+++ lib/src/controllers/th2_file_edit_move_scale_rotate_element_controller.g.dart
@@ -1,0 +1,85 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'th2_file_edit_move_scale_rotate_element_controller.dart';
+
+// **************************************************************************
+// StoreGenerator
+// **************************************************************************
+
+// ignore_for_file: non_constant_identifier_names, unnecessary_brace_in_string_interps, unnecessary_lambdas, prefer_expression_function_bodies, lines_longer_than_80_chars, avoid_as, avoid_annotating_with_dynamic, no_leading_underscores_for_local_identifiers
+
+mixin _$TH2FileEditMoveScaleRotateElementController
+    on TH2FileEditMoveScaleRotateElementControllerBase, Store {
+  late final _$TH2FileEditMoveScaleRotateElementControllerBaseActionController =
+      ActionController(
+        name: 'TH2FileEditMoveScaleRotateElementControllerBase',
+        context: context,
+      );
+
+  @override
+  void moveSelectedElementsToCanvasCoordinates(
+    Offset canvasCoordinatesFinalPosition,
+  ) {
+    final _$actionInfo =
+        _$TH2FileEditMoveScaleRotateElementControllerBaseActionController
+            .startAction(
+              name:
+                  'TH2FileEditMoveScaleRotateElementControllerBase.moveSelectedElementsToCanvasCoordinates',
+            );
+    try {
+      return super.moveSelectedElementsToCanvasCoordinates(
+        canvasCoordinatesFinalPosition,
+      );
+    } finally {
+      _$TH2FileEditMoveScaleRotateElementControllerBaseActionController
+          .endAction(_$actionInfo);
+    }
+  }
+
+  @override
+  void moveSelectedEndControlPointsToCanvasCoordinates(
+    Offset canvasCoordinatesFinalPosition,
+  ) {
+    final _$actionInfo =
+        _$TH2FileEditMoveScaleRotateElementControllerBaseActionController
+            .startAction(
+              name:
+                  'TH2FileEditMoveScaleRotateElementControllerBase.moveSelectedEndControlPointsToCanvasCoordinates',
+            );
+    try {
+      return super.moveSelectedEndControlPointsToCanvasCoordinates(
+        canvasCoordinatesFinalPosition,
+      );
+    } finally {
+      _$TH2FileEditMoveScaleRotateElementControllerBaseActionController
+          .endAction(_$actionInfo);
+    }
+  }
+
+  @override
+  void moveSelectedControlPointToCanvasCoordinates(
+    Offset canvasCoordinatesFinalPosition,
+  ) {
+    final _$actionInfo =
+        _$TH2FileEditMoveScaleRotateElementControllerBaseActionController
+            .startAction(
+              name:
+                  'TH2FileEditMoveScaleRotateElementControllerBase.moveSelectedControlPointToCanvasCoordinates',
+            );
+    try {
+      return super.moveSelectedControlPointToCanvasCoordinates(
+        canvasCoordinatesFinalPosition,
+      );
+    } finally {
+      _$TH2FileEditMoveScaleRotateElementControllerBaseActionController
+          .endAction(_$actionInfo);
+    }
+  }
+
+  @override
+  String toString() {
+    return '''
+
+    ''';
+  }
+}